requirements.txt
aiortc==1.2.0
typing_extensions==3.7.4.3
opencv-python-headless
av==8.0.3
numpy==1.19.5
streamlit==0.83.0
matplotlib==3.4.1
requests==2.22.0
Streamlit
Pillow
Roboflow
pickles
streamlit_webrtc
python-dotenv
